fix: separate invariant error to domain entities, change it to general error

+++ src/Applications/use_case/_test/DeleteCommentUseCase.test.js
@@ -1,37 +1,8 @@
 const DeleteCommentUseCase = require('../DeleteCommentUseCase');
-const InvariantError = require('../../../Commons/exceptions/InvariantError');
+const ThreadRepository = require('../../../Domains/threads/ThreadRepository');
+const CommentRepository = require('../../../Domains/comments/CommentRepository');
 
 describe('DeleteCommentUseCase', () => {
-  it('should throw InvariantError when payload does not contain needed property', async () => {
-    // Arrange
-    const useCasePayload = {
-      // thread tidak ada
-      commentId: 'comment-123',
-      owner: 'user-123',
-    };
-
-    const deleteCommentUseCase = new DeleteCommentUseCase({});
-
-    // Action & Assert
-    await expect(deleteCommentUseCase.execute(useCasePayload))
-      .rejects
-      .toThrow(InvariantError);
-  });
-
-  it('should throw InvariantError when payload has wrong data type', async () => {
-    // Arrange
-    const useCasePayload = {
-      commentId: 123, // content bukan string
-    };
-
-    const deleteCommentUseCase = new DeleteCommentUseCase({});
-
-    // Action & Assert
-    await expect(deleteCommentUseCase.execute(useCasePayload))
-      .rejects
-      .toThrow(InvariantError);
-  });
-
   it('should orchestrate delete comment action correctly', async () => {
     // Arrange
     const useCasePayload = {
@@ -41,14 +12,12 @@ describe('DeleteCommentUseCase', () => {
     };
 
     // mock
-    const mockThreadRepository = {
-      verifyThreadExists: jest.fn(() => Promise.resolve()),
-    };
+    const mockThreadRepository = new ThreadRepository();
+    mockThreadRepository.verifyThreadExists = jest.fn(() => Promise.resolve());
 
-    const mockCommentRepository = {
-      verifyCommentOwner: jest.fn(() => Promise.resolve()),
-      deleteComment: jest.fn(() => Promise.resolve()),
-    };
+    const mockCommentRepository = new CommentRepository();
+    mockCommentRepository.verifyCommentOwner = jest.fn(() => Promise.resolve());
+    mockCommentRepository.deleteComment = jest.fn(() => Promise.resolve());
 
     // instance deleteCommentUseCase dengan mock repository
     const deleteCommentUseCase = new DeleteCommentUseCase({
